remove docker-java-transport-common for now
