+++ docker-java-transport-httpclient5/pom.xml
@@ -35,9 +35,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>docker-java-transport-common</artifactId>
-			<version>${project.version}</version>
+			<groupId>net.java.dev.jna</groupId>
+			<artifactId>jna-platform</artifactId>
+			<version>5.5.0</version>
 		</dependency>
 	</dependencies>
