+++ pom.xml
@@ -89,7 +89,6 @@
 	<modules>
 		<module>docker-java-api</module>
 		<module>docker-java-core</module>
-		<module>docker-java-transport-common</module>
 		<module>docker-java-transport-netty</module>
 		<module>docker-java-transport-jersey</module>
 		<module>docker-java-transport-okhttp</module>
